+++ air/src/air/mod.rs
@@ -6,10 +6,7 @@
 use alloc::{collections::BTreeMap, vec::Vec};
 
 use crypto::{RandomCoin, RandomCoinError};
-use math::{
-    fft, get_power_series, get_power_series_with_offset, ExtensibleField, ExtensionOf,
-    FieldElement, StarkField, ToElements,
-};
+use math::{fft, ExtensibleField, ExtensionOf, FieldElement, StarkField, ToElements};
 
 use crate::{BatchingMethod, ProofOptions};
 
@@ -581,62 +578,17 @@ pub trait Air: Send + Sync {
         E: FieldElement<BaseField = Self::BaseField>,
         R: RandomCoin<BaseField = Self::BaseField>,
     {
-        let mut t_coefficients = Vec::new();
         match self.context().options.deep_poly_batching_method() {
-            BatchingMethod::Linear => {
-                for _ in 0..self.trace_info().width() {
-                    t_coefficients.push(public_coin.draw()?);
-                }
-
-                let mut c_coefficients = Vec::new();
-                for _ in 0..self.context().num_constraint_composition_columns() {
-                    c_coefficients.push(public_coin.draw()?);
-                }
-
-                let lagrange_cc = if self.context().has_lagrange_kernel_aux_column() {
-                    Some(public_coin.draw()?)
-                } else {
-                    None
-                };
-
-                Ok(DeepCompositionCoefficients {
-                    trace: t_coefficients,
-                    constraints: c_coefficients,
-                    lagrange: lagrange_cc,
-                })
-            },
-            BatchingMethod::Algebraic => {
-                let mut t_coefficients = Vec::new();
-                let alpha: E = public_coin.draw()?;
-                t_coefficients
-                    .extend_from_slice(&get_power_series(alpha, self.trace_info().width()));
-
-                let mut c_coefficients = Vec::new();
-
-                let alpha_pow_trace_width = alpha.exp(((self.trace_info().width()) as u32).into());
-                c_coefficients.extend_from_slice(&get_power_series_with_offset(
-                    alpha,
-                    alpha_pow_trace_width,
-                    self.context().num_constraint_composition_columns(),
-                ));
-
-                let lagrange_cc = if self.context().has_lagrange_kernel_aux_column() {
-                    Some(
-                        *c_coefficients
-                            .last()
-                            .expect("there should be at least one constraint coefficient")
-                            * alpha,
-                    )
-                } else {
-                    None
-                };
-
-                Ok(DeepCompositionCoefficients {
-                    trace: t_coefficients,
-                    constraints: c_coefficients,
-                    lagrange: lagrange_cc,
-                })
-            },
+            BatchingMethod::Linear => DeepCompositionCoefficients::draw_linear(
+                public_coin,
+                self.trace_info().width(),
+                self.context().num_constraint_composition_columns(),
+            ),
+            BatchingMethod::Algebraic => DeepCompositionCoefficients::draw_algebraic(
+                public_coin,
+                self.trace_info().width(),
+                self.context().num_constraint_composition_columns(),
+            ),
         }
     }
 }
